feat: add Farcaster avatar as resolver
