+++ src/constants.json
@@ -3,7 +3,7 @@
   "ttl": 86400,
   "shortTtl": 3600,
   "resolvers": {
-    "avatar": ["lens", "ens", "snapshot", "starknet"],
+    "avatar": ["farcaster", "lens", "ens", "snapshot", "starknet"],
     "token": ["trustwallet", "zapper"],
     "space": ["space"],
     "space-sx": ["space-sx"],
