+++ test/integration/resolvers/farcaster.test.ts
@@ -1,0 +1,25 @@
+import resolvers from '../../../src/resolvers';
+
+describe('resolvers', () => {
+  describe('farcaster', () => {
+    it('should return false if missing', async () => {
+      const result = await resolvers.farcaster('0x556B14CbdA79A36dC33FcD461a04A5BCb5dC2A70');
+
+      expect(result).toBe(false);
+    });
+
+    it('should resolve with handle', async () => {
+      const result = await resolvers.farcaster('fabien.lens');
+
+      expect(result).toBeInstanceOf(Buffer);
+      expect(result.length).toBeGreaterThan(1000);
+    });
+
+    it('should resolve with address', async () => {
+      const result = await resolvers.farcaster('0xeF8305E140ac520225DAf050e2f71d5fBcC543e7');
+
+      expect(result).toBeInstanceOf(Buffer);
+      expect(result.length).toBeGreaterThan(1000);
+    });
+  });
+});
